--- a/typed-racket-lib/info.rkt
+++ b/typed-racket-lib/info.rkt
@@ -11,4 +11,4 @@
 
 (define pkg-authors '(samth stamourv))
 
-(define version "1.8")
+(define version "1.9")