--- a/typed-racket-lib/typed-racket/base-env/base-types-extra.rkt
+++ b/typed-racket-lib/typed-racket/base-env/base-types-extra.rkt
@@ -19,7 +19,7 @@
 
 ;; special type names that are not bound to particular types
 (define-other-types
-  -> ->* case-> U Union ∩ Intersection Rec All Opaque Vector
+  -> ->* case-> U Union ∩ Intersection Rec All Opaque Immutable-Vector Mutable-Vector Vector
   Parameterof List List* Class Object Row Unit Values AnyValues Instance Refinement
   pred Struct Struct-Type Prefab PrefabTop Distinction Sequenceof Refine)