--- a/typed-racket-lib/typed-racket/base-env/base-types.rkt
+++ b/typed-racket-lib/typed-racket/base-env/base-types.rkt
@@ -125,6 +125,7 @@
 [ChannelTop -ChannelTop]
 [Async-ChannelTop -Async-ChannelTop]
 [VectorTop -VectorTop]
+[Mutable-VectorTop -Mutable-VectorTop]
 [HashTableTop -HashTableTop]
 [MPairTop -MPairTop]
 [Thread-CellTop -ThreadCellTop]
@@ -171,6 +172,8 @@
 
 [Listof -Listof]
 [Vectorof (-poly (a) (make-Vector a))]
+[Immutable-Vectorof (-poly (a) (-ivec a))]
+[Mutable-Vectorof (-poly (a) (-mvec a))]
 [FlVector -FlVector]
 [ExtFlVector -ExtFlVector]
 [FxVector -FxVector]