--- a/typed-racket-test/fail/for-vector-bad-ann.rkt
+++ b/typed-racket-test/fail/for-vector-bad-ann.rkt
@@ -1,0 +1,5 @@
+#;
+(exn-pred 3)
+#lang typed/racket
+
+(for/vector : (Immutable-Vectorof Any) ([x (in-list '())]) x)