--- a/typed-racket-test/fail/heterogeneous-immutable-vector-bad-ann.rkt
+++ b/typed-racket-test/fail/heterogeneous-immutable-vector-bad-ann.rkt
@@ -1,0 +1,5 @@
+#;
+(exn-pred #rx"given: \\(Immutable-Vector One\\)")
+#lang typed/racket
+
+(ann '#(1) (Mutable-Vector Real))