New types:
- Immutable-Vector (covariant)
- Mutable-Vector (invariant)
- Immutable-Vectorof (covariant)
- Mutable-Vectorof (invariant)
- Mutable-VectorTop (similar to Mutable-HashTableTop)

The old vector types are now unions of these new types.

Also:
- more specific _return_ types for some built-in functions (e.g. `vector-append`)
- `immutable?` distinguishes mutable & immutable vectors
- the contract for a `Vector` or `Vectorof` should be the same as before,
  even though these types are now unions

Current TODOs:
- [x] travis
- [x] think about adding more tests
- [x] clean up the "union deduplicating" in `type-contract.rkt`, it's right now an awkward implementation of a simple idea
- [X] "oops" below, revert the condition and make sure tests pass. EDIT: reverting the condition breaks some `typecheck-tests`, not sure why.
- [X] "oops 2", add a test for this bug & fix it
- [X] "oops 3" ... try simplifying the code & see if tests pass?
- [x] make sure HVecs are subtypes of VectorTop
- [x] make sure VectorTop generates a flat contract (try compiling some packages)
  - found errors in ~`plot`~ (fixed) and `scv`
- [x] sync with fixed https://github.com/racket/racket/issues/1745
- [x] should other `intersect`s in `tc-literal.rkt` be `resolve`s?
  - removed all these intersects; expected type isn't relevant when the type is immutable (because immutable types have subtyping)
- [X] change `type-contract-generalize` interface
  - incorrect to use just one `type->contract` continuation because ignores typed-side/untyped-side distinction
  - and try to make that function overall simpler
  - and should move this code to the static-contracts optimizer
- [x] fix #598 , move tests for `type-contract-generalize` to static-contracts-optimize
- [x] add history annotations
- [X] do we really need travis_wait? (why is typechecking slower) --- NO, do not accept the travis_wait, make this PR faster instead
- [x] resolve #743 (EDIT: resolved the part that matters for this PR)