@bennn 2017-07-10
Problem: I tried giving the result of `for/vector` the type `(Mutable-Vector ....)`, but the implementation of `for/vector` uses a `set!` and so generalizes my `Mutable-Vector` into a `Vector`.

Is there a way to keep the `set!`, but prevent it from generalizing? ([Here](https://github.com/racket/typed-racket/pull/575/files#diff-72cf4e4479067e30d7f7287eeb580300R821) is the `set!`)