@samth 2017-07-10
To fix the issue with `set!`, you need to add an annotation to those variables at the point of definition.