@bennn 2017-07-12
I don't know a simpler way to put a mutable vector into a syntax object.

For immutable vectors:
```
(immutable? (syntax-e #'#(1 2)))
;; ==> #f, seems wrong
```