@bennn 2017-07-12
@pnwamk  the current travis failure is from `with-integer-linear-arith2.rkt`, basically:

```
(define-type (V8 A) (Refine [v : (Vectorof A)] (= 8 (vector-length v))))
(ann (vector 0 1 2 3 4 5 6 7) (V8 Any))
```

should I look at `types/subtype.rkt` to fix this, or do you think the problem might be somewhere else?


I think it's probably subtyping because this code works:
```
(define-type (V8 A) (Refine [v : (Vectorof A)] (= 8 (vector-length v))))
(define-type (MV8 A) (Refine [v : (Mutable-Vectorof A)] (= 8 (vector-length v))))
(ann (ann (vector 0 1 2 3 4 5 6 7) (MV8 Any)) (V8 Any)) 
```
but I figure I might as well ask, because I'm not planning to work on this until later tonight.