@pnwamk 2017-07-12
@bennn Played around a little with the issue, this seemed to fix it:

https://github.com/pnwamk/typed-racket/commit/d1f671588e48a6da5ddaff56196768441289bbc6

I haven't tested it more broadly, though.