@pnwamk 2017-07-12
It's too late over here for me to think this through all the way at the moment, but I wonder if we need a similar match expander (that matches on the unions of the various kinds) for hash tables as well.