@bennn 2017-07-12
Thanks for looking.

I don't know about the match expander. I think the I/M distinction is useful and don't want to start hiding it internally.

I'll look to see where the match expander helped fix this problem. (Hoping its just 1 or 2 places that need to special-case.)