@bennn 2017-07-13
Ok fixed that test. There was a bug in `tc-literal.rkt`, I was intersecting the expected type with the wrong vector type.

(Next up: the optimizer, which I think definitely needs to recognize `(U IVec MVec)`)