@pnwamk 2017-07-13
One small (possibly entirely unnecessary) note -- if we're going to bother having match expanders like `Vector:` (and I think they are occasionally useful), I think they need to match on the union case. i.e. if `Vector:` intuitively means as long as its "either kind of vector", I think we'll be accidentally introducing bugs if it omits the union case.