@bennn 2017-07-19
short question: `(∩ (Immutable-Vectorof Any) (Sequenceof Real))` should be `(Immutable-Vectorof Real)`, right?

- - - 

story: `(ann #(1/2 1/2) (Sequenceof Real))` no longer typechecks, because `tc-literal.rkt`:
- intersects the expected type with `VectorTop`
- giving the expected type `(U (∩ (IVof Any) (Seqof R)) (∩ MVTop (Seqof R))`
- which is a union where exactly 1 type `t` is such that `(mask t) = mask:immutable-vector`
- so `tc-literal` loops with the new expected type `(∩ (IVof Any) (Seqof R))`
- and intersects that type with `VectorTop`
- giving the expected type `(Immutable-Vectorof Any)`
- so the program no longer type checks

I think either the intersection or mask-checking is wrong.