@samth 2017-07-19
That mask check is definitely right.