@pnwamk 2017-07-19
would it be better to use the non-additive `restrict` instead of `intersect` there? (i.e. `restrict` with `VectorTop` instead of `intersect`)