@pnwamk 2017-07-19
what is `raw-mask`? I don't see it on quick repo searching and C-f of the PR diff.