@bennn 2017-07-19
I got `raw-mask` from the [definition of `mask`](https://github.com/racket/typed-racket/blob/master/typed-racket-lib/typed-racket/rep/rep-utils.rkt#L195). But it's not in this PR.

Using `restrict` is a good idea --- after that change, the `(ann #(1/2 ....))` and all typecheck-tests pass.