@pnwamk 2017-07-19
if that is the right solution, it might mean the other uses of intersect in that file should be restrict as well... ?