@bennn 2017-07-19
Maybe ... I'll try to adapt the Vector programs into something that points out a flaw in other intersects. But if I don't find any programs that fail, I'll leave the `intersect` alone in this PR.