@bennn 2018-04-01
Here's the results. I'll try to speed up `fsm` and `dungeon` this week.

(Key: `I` = this pull request, `M` = master, `Y` = with contracts, `N` = without contracts)

 program                     | I Y    | M Y    | I N    | M N    |
------------------------- | -------- | -------- | -------- | --------
 schml-specify-rep       | 20.80s | 19.12s | 14.98s | 14.10s 
 schml-interp-casts-help | 45.01s | 42.79s | 30.14s | 31.25s 
 parser                  |  4.32s |  4.30s |  3.60s |  3.55s 
 old-metrics             |  5.23s |  5.24s |  4.25s |  4.16s 
 new-metrics             |  8.50s |  8.38s |  5.65s |  5.63s 
 math-flonum             |  7.94s |  7.91s |  6.07s |  5.82s 
 fsm                     | 11.17s |  9.73s |  8.16s |  7.99s 
 forth                   | 10.18s |  9.49s |  7.86s |  7.83s 
 dungeon                 | 36.69s | 17.37s | 26.79s | 14.09s 
 acquire                 | 36.24s | 35.40s | 26.52s | 27.21s